"test it publish metrics to ingestion telemetry endpoint" requires env variable to be ran
